Affilaites

+++ dapps/shop/backend/scripts/printful/getProduct.js
@@ -8,10 +8,14 @@ async function getProduct({ apiAuth, OutputDir, id }) {
 
   const productId = json.result.sync_variants[0].product.product_id
   const productPath = `${OutputDir}/data-printful/product-${productId}.json`
-  const productJson = await get(`/products/${productId}`, { auth: apiAuth })
-  fs.writeFileSync(productPath, JSON.stringify(productJson.result, null, 2))
 
-  console.log(`Wrote sync product ${id}, product ${productId}`)
+  if (!fs.existsSync(productPath)) {
+    const productJson = await get(`/products/${productId}`, { auth: apiAuth })
+    fs.writeFileSync(productPath, JSON.stringify(productJson.result, null, 2))
+    console.log(`Wrote sync product ${id}, product ${productId}`)
+  } else {
+    console.log(`Wrote sync product ${id}`)
+  }
 }
 
 module.exports = getProduct
